+++ packages/library-base/src/jvm/kotlin/io/realm/internal/platform/SystemUtils.kt
@@ -1,15 +1,15 @@
 package io.realm.internal.platform
 
 @Suppress("MayBeConst") // Cannot make expect/actual const
-actual val RUNTIME: String = "JVM"
+public actual val RUNTIME: String = "JVM"
 
-actual fun threadId(): ULong {
+public actual fun threadId(): ULong {
     return Thread.currentThread().id.toULong()
 }
 
-actual fun <T> T.freeze(): T = this
+public actual fun <T> T.freeze(): T = this
 
-actual val <T> T.isFrozen: Boolean
+public actual val <T> T.isFrozen: Boolean
     get() = false
 
-actual fun Any.ensureNeverFrozen() {}
+public actual fun Any.ensureNeverFrozen() {}